sample/pyproject.toml
[project]
name = "sample"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "okmtdev", email = "okmtdev@gmail.com" }
]
dependencies = [
    "python-pptx>=1.0.2",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
